Refactor eats test suite to run tests on deployed eirini
When running on a deployed eirini we do not need to run local binaries.
Instead, we are requesting the endpoint of the opi that has been already
deployed

The `USE_DEPLOYED_EIRINI` env variable determines whether we are running
agains the deployed eirini or not.

There are 2 tests still failing when you use USE_DEPLOYED_EIRINI because
they rely on a fake cloud controller being started
(Might be fixed with this: https://www.pivotaltracker.com/story/show/174193327).

+++ integration/eats/eats_suite_test.go
@@ -76,23 +76,37 @@ var _ = SynchronizedAfterSuite(func() {}, func() {
 var _ = BeforeEach(func() {
 	fixture.SetUp()
 
-	localhostCertPath, localhostKeyPath = util.GenerateKeyPair("localhost")
-
 	var err error
-	httpClient, err = makeTestHTTPClient(localhostCertPath, localhostKeyPath)
-	Expect(err).ToNot(HaveOccurred())
+	if util.IsUsingDeployedEirini() {
+		localhostCertPath, localhostKeyPath = util.DownloadEiriniCertificates()
 
-	opiSession, opiConfig, opiURL = runOpi(localhostCertPath, localhostKeyPath)
-	waitOpiReady(httpClient, opiURL)
+		httpClient, err = makeTestHTTPClient(localhostCertPath, localhostKeyPath)
+		Expect(err).ToNot(HaveOccurred())
+
+		opiHost := util.KubeCtl("get svc -n eirini-core eirini-external -o jsonpath=\"{.status.loadBalancer.ingress[0].ip}\"")
+		opiURL = fmt.Sprintf("https://%s:8085", opiHost)
+	} else {
+		localhostCertPath, localhostKeyPath = util.GenerateKeyPair("localhost")
+
+		opiSession, opiConfig, opiURL = runOpi(localhostCertPath, localhostKeyPath)
+
+		httpClient, err = makeTestHTTPClient(localhostCertPath, localhostKeyPath)
+		Expect(err).ToNot(HaveOccurred())
+
+		waitOpiReady(httpClient, opiURL)
+	}
 })
 
 var _ = AfterEach(func() {
 	fixture.TearDown()
 
-	if opiSession != nil {
-		opiSession.Kill()
+	if !util.IsUsingDeployedEirini() {
+		if opiSession != nil {
+			opiSession.Kill()
+		}
+		Expect(os.Remove(opiConfig)).To(Succeed())
 	}
-	Expect(os.Remove(opiConfig)).To(Succeed())
+
 	Expect(os.Remove(localhostCertPath)).To(Succeed())
 	Expect(os.Remove(localhostKeyPath)).To(Succeed())
 })
